Startup test added to LEDstream (Arduino), shutdown added to Adalight (Processing)
Upon initial power-up or reset, and each time the serial port is opened,
the LEDs will briefly flash through red, green and blue, then off. This
can be used to diagnose whether the wiring between Arduino and LEDs is
correct, and to spot any breaks in the strand. Unrelated, Adalight (the
Processing code) now issues an all-off state to the LEDs on most (but
not all) exit cases.