Write to LED indices instead of raw pointer
The original implementation violates strict aliasing. It works most of the time but can potentially cause issues.

Note that the LED index counters are 32-bit instead of 16-bit because the LED count provided by the checksum has a +1, so it will not fit in a 16-bit variable.

Similarly, we need to track both the current LED index and the remaining LEDs to write because the size of the LED strip in the code may not match the size of the LED strip as set by the data protocol.